Giao diện cơ bản

+++ src/main/java/com/t3h/e_commerce/repository/OrderItemRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderItemRepository extends JpaRepository<OrderItemEntity, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItemEntity, Long> {
 }
+
